// src/react-turbo-select/utils/classNames.d.ts
export declare const CLASS_NAMES: {
    CONTROLS_CONTAINER: string;
    DROP_DOWN_ICON: string;
    REMOVE_ICON: string;
    MENU: string;
    MENU_OPEN: string;
    MENU_CLOSE: string;
    OPTION: string;
    OPTION_SELECTED: string;
    OPTION_SELECTED_DARK: string;
    INPUT: string;
    MULTIPLE_MENU: string;
    TAG: string;
    REMOVE_TAG_ICON: string;
};